refractored the daily schedule widget and the todolist widget into their own file
